+++ src/views/NormalBattle/ShowVertexTerrain/ShowVertexTerrain.tsx
@@ -1,0 +1,9 @@
+export default ShowVertexTerrain;
+
+function ShowVertexTerrain() {
+  return (
+    <div>
+      <div>ShowVertexTerrain</div>
+    </div>
+  );
+}