+++ src/views/NormalBattle/ShowVertexTerrain/index.ts
@@ -1,0 +1,2 @@
+import ShowVertexTerrain from "./ShowVertexTerrain";
+export default ShowVertexTerrain;