+++ src/views/NormalBattle/ShowVertexTypography/ShowVertexTypography.tsx
@@ -1,9 +1,0 @@
-export default ShowVertexTypography;
-
-function ShowVertexTypography() {
-  return (
-    <div>
-      <div>ShowVertexTypography</div>
-    </div>
-  );
-}