+++ src/views/NormalBattle/ShowVertexTypography/index.ts
@@ -1,2 +1,0 @@
-import ShowVertexTypography from "./ShowVertexTypography";
-export default ShowVertexTypography;